Refine rig artifact index structure
